auth-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic[email]==2.6.1
alembic==1.13.2
